Change the oauth scope to `*`
Rather than specify all of the scopes, we use the
`*` scope which expands to all public scopes while
still leaving the ability to specify additional
scopes.

This makes us more consistent with what is done
on the redesign and avoids issues when additional
scopes are added to the system but not to all
clients.